combinatory approach in C#, with recursive delegates
If one is to use the combinatory approach in C#, this is the right
way to do it. The other ways are distinctly inferior and should
probably be reserved just for exposition.

Delegates can be recursive (i.e., they are in scope in their own
definitions and they can appear in, or even as, their own parameter
or return types). So this gives a completely type-safe way to make
function objects (delegate instances) that accept themselves as
arguments.

The mindset I had to break out of was always trying to use
System.Func (or System.Action). Clearly this is a case where that
is not a fruitful approach.

This should carry over to Java as well. Java doesn't have
delegates, but instead of defining delegate types, one defines
functional interfaces, and an interface can declare methods where
the interface itself appears in (or as) a parameter or return type.